+++ pnpm/crates/deps-inspection-peers/src/tests.rs
