+++ pnpm/crates/package-manager/Cargo.toml
@@ -36,6 +36,7 @@ pnpm-modules-yaml                      = { workspace = true }
 pnpm-network                           = { workspace = true }
 pnpm-config                            = { workspace = true }
 pnpm-config-parse-overrides            = { workspace = true }
+pnpm-deps-inspection-peers             = { workspace = true }
 pnpm-graph-hasher                      = { workspace = true }
 pnpm-hooks                             = { workspace = true }
 pnpm-package-manifest                  = { workspace = true }
