+++ pnpm/crates/package-manager/src/install.rs
@@ -589,6 +589,15 @@ pub enum InstallError {
         package_names: Vec<String>,
     },
 
+    /// pnpm's `ERR_PNPM_PEER_DEP_ISSUES`: with `strictPeerDependencies`
+    /// on, an install whose resolution left unmet peers behind fails
+    /// once the artifacts are written, the same way `IgnoredBuilds`
+    /// does — the tree is installed, and the run reports the verdict on
+    /// it.
+    #[display("Unmet peer dependencies\n\n{rendered}")]
+    #[diagnostic(code(ERR_PNPM_PEER_DEP_ISSUES), help("{hints}"))]
+    PeerDependencyIssues { rendered: String, hints: String },
+
     /// A custom resolver hook failed (loading the pnpmfile's resolvers
     /// or running `shouldRefreshResolution`) while deciding whether the
     /// frozen-path optimization may run. A throwing hook aborts the
